Fixing syntax error with multiple brackets in view
In play views, if a phExpr syntax region had extra brackets in then the
syntax highlighting did not work correctly.

e.g.

@for((foo, number) <- fooList.view.zipWithIndex) {
    <div>@(foo) is number @(number)</div>
}

the extra brackets around f00,number in the for loop would cause
problems.

By changing the phExpr region definition so that it can contain other
phExpr regions this seems to fix the problem.